+++ src/pages/JefeDepartamento/DeudasYDescargas/MantenimientoProcesos/ItemProcesoActual.js
@@ -6,7 +6,7 @@ import React from 'react'
 export default function ItemProcesoActual() {
     return (
         <div>
-            
+            :D
         </div>
     )
 }
